Fixed code to work with latest version of testify/assert

+++ ch08/q1_test.go
@@ -20,12 +20,12 @@ func FibRecursive(n uint) uint {
 }
 
 func TestFibRecursive(t *testing.T) {
-	assert.Equal(t, 1, FibRecursive(0))
-	assert.Equal(t, 2, FibRecursive(1))
-	assert.Equal(t, 3, FibRecursive(2))
-	assert.Equal(t, 5, FibRecursive(3))
-	assert.Equal(t, 8, FibRecursive(4))
-	assert.Equal(t, 13, FibRecursive(5))
+	assert.Equal(t, uint(1), FibRecursive(0))
+	assert.Equal(t, uint(2), FibRecursive(1))
+	assert.Equal(t, uint(3), FibRecursive(2))
+	assert.Equal(t, uint(5), FibRecursive(3))
+	assert.Equal(t, uint(8), FibRecursive(4))
+	assert.Equal(t, uint(13), FibRecursive(5))
 }
 
 func FibIterative(n uint) (sum uint) {
@@ -38,10 +38,10 @@ func FibIterative(n uint) (sum uint) {
 }
 
 func TestFibIterative(t *testing.T) {
-	assert.Equal(t, 1, FibIterative(0))
-	assert.Equal(t, 2, FibIterative(1))
-	assert.Equal(t, 3, FibIterative(2))
-	assert.Equal(t, 5, FibIterative(3))
-	assert.Equal(t, 8, FibIterative(4))
-	assert.Equal(t, 13, FibIterative(5))
+	assert.Equal(t, uint(1), FibIterative(0))
+	assert.Equal(t, uint(2), FibIterative(1))
+	assert.Equal(t, uint(3), FibIterative(2))
+	assert.Equal(t, uint(5), FibIterative(3))
+	assert.Equal(t, uint(8), FibIterative(4))
+	assert.Equal(t, uint(13), FibIterative(5))
 }
